OSCORE appendix-B.2 tests with IDContext!=null

+++ leshan-server-cf/src/main/java/org/eclipse/leshan/server/californium/OscoreContextCleaner.java
@@ -74,6 +74,9 @@ public class OscoreContextCleaner implements RegistrationListener, SecurityStore
     private void removeContext(byte[] rid) {
         OSCoreCtx context = oscoreCtxDB.getContext(rid);
         if (context != null)
-            oscoreCtxDB.removeContext(context);
+            // JL test
+            if (!oscoreCtxDB.getContext(rid).getContextRederivationEnabled()) {
+                oscoreCtxDB.removeContext(context);
+            }
     }
 }
